requirements.txt
apache-beam[gcp]==2.55.0
google-cloud-pubsub==2.21.1
google-cloud-bigquery==3.17.1
google-cloud-storage==2.16.0

pytest==8.1.1
pytest-cov==5.0.0
